Implement new Projection. Move most Overlays to new Projection.

+++ OpenStreetMapViewer/src/org/osmdroid/samples/SampleExtensive.java
@@ -75,8 +75,9 @@ public class SampleExtensive extends SampleMapActivity implements OpenStreetMapC
 			this.mOsmv.getOverlays().add(mScaleBarOverlay);
 			// Scale bar tries to draw as 1-inch, so to put it in the top center, set x offset to
 			// half screen width, minus half an inch.
-			this.mScaleBarOverlay.setScaleBarOffset(getResources().getDisplayMetrics().widthPixels
-					/ 2 - getResources().getDisplayMetrics().xdpi / 2, 10);
+			this.mScaleBarOverlay.setScaleBarOffset(
+					(int) (getResources().getDisplayMetrics().widthPixels / 2 - getResources()
+							.getDisplayMetrics().xdpi / 2), 10);
 		}
 
 		/* SingleLocation-Overlay */
